+++ src/containers/ConfirmationContainer/components/ConfirmModal/ConfirmModal.vue
@@ -93,10 +93,9 @@
 
 <script>
 import AddressBlock from '../AddressBlock';
-import BigNumber from 'bignumber.js';
 import { mapGetters } from 'vuex';
-import store from 'store';
 import StandardButton from '@/components/Buttons/StandardButton';
+import parseTokensData from '@/helpers/parseTokensData.js';
 
 export default {
   components: {
@@ -180,71 +179,30 @@ export default {
   },
   watch: {
     data(newVal) {
-      this.parseData(newVal);
+      parseTokensData(
+        newVal,
+        this.to,
+        this.web3,
+        this.network.type.tokens,
+        this.network.type.name
+      );
     }
   },
   mounted() {
     if (this.data !== '0x') {
-      this.parseData(this.data);
+      parseTokensData(
+        this.data,
+        this.to,
+        this.web3,
+        this.network.type.tokens,
+        this.network.type.name
+      );
     }
   },
   methods: {
     sendTx() {
       if (this.signedTx !== '') {
         this.confirmSendTx();
-      }
-    },
-    async parseData(data) {
-      const web3 = this.web3;
-      const networkToken = this.network.type.tokens;
-
-      let token = networkToken.find(el => {
-        return el.address.toLowerCase() === this.to.toLowerCase();
-      });
-      if (!token) {
-        const customStore = store.get('customTokens');
-        if (
-          customStore !== undefined &&
-          customStore[this.network.type.name] !== undefined &&
-          customStore[this.network.type.name].length
-        ) {
-          token = customStore[this.network.type.name].find(el => {
-            return el.address.toLowerCase() === this.to.toLowerCase();
-          });
-        }
-      }
-      const jsonInterface = {
-        constant: false,
-        inputs: [
-          { name: '_to', type: 'address' },
-          { name: '_amount', type: 'uint256' }
-        ],
-        name: 'transfer',
-        outputs: [{ name: '', type: 'bool' }],
-        payable: false,
-        stateMutability: 'nonpayable',
-        type: 'function'
-      };
-      const transferFuncSig = web3.eth.abi.encodeFunctionSignature(
-        jsonInterface
-      );
-      this.tokenTransferTo = '';
-      this.tokenTransferVal = '';
-      this.tokenSymbol = '';
-      if (data.substr(0, 10) === transferFuncSig) {
-        const params = web3.eth.abi.decodeParameters(
-          ['address', 'uint256'],
-          `${data.substr(10)}`
-        );
-        const value = new BigNumber(params[1]);
-        this.tokenTransferTo = params[0];
-        this.tokenTransferVal = token
-          ? value
-              .div(new BigNumber(10).pow(token.decimals))
-              .toFixed()
-              .toString()
-          : value.toString();
-        this.tokenSymbol = token ? token.symbol : 'Unidentified Token';
       }
     }
   }